refactor(split/t1-b1): mechanically split parity test target
