+++ crates/assay-core/tests/parity.rs
@@ -11,926 +11,22 @@
 //! CI gate:
 //!   Any parity failure is a BLOCKER for release.
 
-use assay_core::policy_engine::{evaluate_tool_args, VerdictStatus};
-use serde::{Deserialize, Serialize};
-
-// ============================================================
-// Core Types
-// ============================================================
-
-/// A single policy check to evaluate
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct PolicyCheck {
-    pub id: String,
-    pub check_type: CheckType,
-    pub params: serde_json::Value,
-}
-
-/// Types of policy checks
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum CheckType {
-    ArgsValid,
-    SequenceValid,
-    ToolBlocklist,
-}
-
-/// Input to a policy check
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct CheckInput {
-    /// Tool being called (for args_valid, blocklist)
-    pub tool_name: Option<String>,
-    /// Arguments to validate
-    pub args: Option<serde_json::Value>,
-    /// Trace of tool calls (for sequence_valid)
-    pub trace: Option<Vec<ToolCall>>,
-}
-
-/// A tool call in a trace
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ToolCall {
-    pub tool_name: String,
-    pub args: serde_json::Value,
-    pub timestamp_ms: u64,
-}
-
-/// Result of a policy check
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct CheckResult {
-    pub check_id: String,
-    pub outcome: Outcome,
-    pub reason: String,
-    /// Canonical hash for comparison
-    pub result_hash: String,
-}
-
-/// Outcome of a check
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum Outcome {
-    Pass,
-    Fail,
-    Error,
-}
-
-// ============================================================
-// Evaluation Engines (Simulated)
-// ============================================================
-
-/// Batch mode evaluation (simulates `assay run`)
-pub mod batch {
-    use super::*;
-
-    pub fn evaluate(check: &PolicyCheck, input: &CheckInput) -> CheckResult {
-        // This simulates the batch evaluation path
-        // In real implementation, this would call assay-core directly
-
-        let (outcome, reason) = match check.check_type {
-            CheckType::ArgsValid => evaluate_args_valid(&check.params, input),
-            CheckType::SequenceValid => evaluate_sequence_valid(&check.params, input),
-            CheckType::ToolBlocklist => evaluate_blocklist(&check.params, input),
-        };
-
-        let result_hash = compute_result_hash(&check.id, &outcome, &reason);
-
-        CheckResult {
-            check_id: check.id.clone(),
-            outcome,
-            reason,
-            result_hash,
-        }
-    }
-
-    fn evaluate_args_valid(params: &serde_json::Value, input: &CheckInput) -> (Outcome, String) {
-        crate::shared::args_valid(params, input)
-    }
-
-    fn evaluate_sequence_valid(
-        params: &serde_json::Value,
-        input: &CheckInput,
-    ) -> (Outcome, String) {
-        crate::shared::sequence_valid(params, input)
-    }
-
-    fn evaluate_blocklist(params: &serde_json::Value, input: &CheckInput) -> (Outcome, String) {
-        crate::shared::blocklist(params, input)
-    }
-}
-
-/// Streaming mode evaluation (simulates `assay-mcp-server`)
-pub mod streaming {
-    use super::*;
-
-    pub fn evaluate(check: &PolicyCheck, input: &CheckInput) -> CheckResult {
-        // This simulates the streaming/preflight evaluation path
-        // In real implementation, this would call the MCP server handler
-
-        // CRITICAL: Must use IDENTICAL logic to batch mode
-        // The only difference should be the execution context, not the logic
-
-        let (outcome, reason) = match check.check_type {
-            CheckType::ArgsValid => evaluate_args_valid(&check.params, input),
-            CheckType::SequenceValid => evaluate_sequence_valid(&check.params, input),
-            CheckType::ToolBlocklist => evaluate_blocklist(&check.params, input),
-        };
-
-        let result_hash = compute_result_hash(&check.id, &outcome, &reason);
-
-        CheckResult {
-            check_id: check.id.clone(),
-            outcome,
-            reason,
-            result_hash,
-        }
-    }
-
-    // These functions MUST be identical to batch mode
-    // In production, both modes should call the same underlying functions
-    // from assay-metrics
-
-    fn evaluate_args_valid(params: &serde_json::Value, input: &CheckInput) -> (Outcome, String) {
-        // Delegate to shared implementation
-        crate::shared::args_valid(params, input)
-    }
-
-    fn evaluate_sequence_valid(
-        params: &serde_json::Value,
-        input: &CheckInput,
-    ) -> (Outcome, String) {
-        crate::shared::sequence_valid(params, input)
-    }
-
-    fn evaluate_blocklist(params: &serde_json::Value, input: &CheckInput) -> (Outcome, String) {
-        crate::shared::blocklist(params, input)
-    }
-}
-
-/// Shared evaluation logic (the "single engine")
-pub mod shared {
-    use super::*;
-
-    pub fn args_valid(params: &serde_json::Value, input: &CheckInput) -> (Outcome, String) {
-        let schema = match params.get("schema") {
-            Some(s) => s,
-            None => return (Outcome::Error, "Config error: schema missing".into()),
-        };
-
-        // Wrap simple schema in tool map for policy_engine
-        let tool_name = input.tool_name.as_deref().unwrap_or("unknown");
-        let policy = serde_json::json!({
-             tool_name: schema
-        });
-
-        let args = match &input.args {
-            Some(a) => a,
-            None => return (Outcome::Error, "No args provided".into()),
-        };
-
-        let verdict = evaluate_tool_args(&policy, tool_name, args);
-
-        match verdict.status {
-            VerdictStatus::Allowed => (Outcome::Pass, "args valid".into()),
-            VerdictStatus::Blocked => {
-                // Map reason codes to test expectations if needed
-                if verdict.reason_code == "E_ARG_SCHEMA" {
-                    // Extract first violation for parity valid reason check
-                    // The test expects "percent 50 exceeds maximum 30" etc.
-                    // The real engine returns structured JSON violations.
-                    // We need to adapt the message to match the mock test cases OR update test cases.
-                    // For V1 integration, ensuring Outcome matches is priority #1.
-                    // The mock test strings are very specific "percent {} exceeds maximum {}".
-                    // Real engine says: "data.percent: 50.0 is greater than the maximum of 30.0" (JSON schema output)
-                    // To pass the STRICT parity test provided by the user (which checks `reason == reason`),
-                    // we just need consistent strings.
-                    // Since *both* Batch and Streaming call *this* wrapper, they will get identical strings.
-                    // So we can return a generic string or the detailed one.
-                    // Let's return the structured details stringified
-                    (
-                        Outcome::Fail,
-                        format!("Schema violation: {}", verdict.details),
-                    )
-                } else if verdict.reason_code == "E_POLICY_MISSING_TOOL" {
-                    (Outcome::Error, "Tool not in policy".into())
-                } else {
-                    (Outcome::Fail, format!("Blocked: {}", verdict.reason_code))
-                }
-            }
-        }
-    }
-
-    pub fn sequence_valid(params: &serde_json::Value, input: &CheckInput) -> (Outcome, String) {
-        let trace = match &input.trace {
-            Some(t) => t,
-            None => return (Outcome::Error, "No trace provided".into()),
-        };
-
-        let tool_names: Vec<&str> = trace.iter().map(|t| t.tool_name.as_str()).collect();
-
-        if let Some(rules) = params.get("rules").and_then(|r| r.as_array()) {
-            for rule in rules {
-                if let Some(rule_type) = rule.get("type").and_then(|t| t.as_str()) {
-                    match rule_type {
-                        "require" => {
-                            if let Some(tool) = rule.get("tool").and_then(|t| t.as_str()) {
-                                if !tool_names.contains(&tool) {
-                                    return (
-                                        Outcome::Fail,
-                                        format!("required tool not called: {}", tool),
-                                    );
-                                }
-                            }
-                        }
-                        "before" => {
-                            let first = rule.get("first").and_then(|t| t.as_str());
-                            let then = rule.get("then").and_then(|t| t.as_str());
-
-                            if let (Some(first), Some(then)) = (first, then) {
-                                let first_idx = tool_names.iter().position(|&t| t == first);
-                                let then_idx = tool_names.iter().position(|&t| t == then);
-
-                                match (first_idx, then_idx) {
-                                    (Some(f), Some(t)) if f >= t => {
-                                        return (
-                                            Outcome::Fail,
-                                            format!("{} must come before {}", first, then),
-                                        );
-                                    }
-                                    (None, Some(_)) => {
-                                        return (
-                                            Outcome::Fail,
-                                            format!("{} must come before {}", first, then),
-                                        );
-                                    }
-                                    _ => {}
-                                }
-                            }
-                        }
-                        _ => {}
-                    }
-                }
-            }
-        }
-
-        (Outcome::Pass, "sequence valid".into())
-    }
-
-    pub fn blocklist(params: &serde_json::Value, input: &CheckInput) -> (Outcome, String) {
-        let tool = match &input.tool_name {
-            Some(t) => t,
-            None => return (Outcome::Error, "No tool_name provided".into()),
-        };
-
-        if let Some(blocked) = params.get("blocked").and_then(|b| b.as_array()) {
-            for b in blocked {
-                if let Some(blocked_name) = b.as_str() {
-                    if tool == blocked_name {
-                        return (Outcome::Fail, format!("tool {} is blocked", tool));
-                    }
-                }
-            }
-        }
-
-        (Outcome::Pass, "tool allowed".into())
-    }
-}
-
-// ============================================================
-// Parity Verification
-// ============================================================
-
-/// Compute a canonical hash for result comparison
-fn compute_result_hash(check_id: &str, outcome: &Outcome, reason: &str) -> String {
-    use std::collections::hash_map::DefaultHasher;
-    use std::hash::{Hash, Hasher};
-
-    let mut hasher = DefaultHasher::new();
-    check_id.hash(&mut hasher);
-    format!("{:?}", outcome).hash(&mut hasher);
-    reason.hash(&mut hasher);
-    format!("{:016x}", hasher.finish())
-}
-
-/// Compare batch and streaming results
-pub fn verify_parity(check: &PolicyCheck, input: &CheckInput) -> ParityResult {
-    let batch_result = batch::evaluate(check, input);
-    let streaming_result = streaming::evaluate(check, input);
-
-    let is_identical = batch_result.outcome == streaming_result.outcome
-        && batch_result.reason == streaming_result.reason;
-
-    ParityResult {
-        check_id: check.id.clone(),
-        batch_result,
-        streaming_result,
-        is_identical,
-    }
-}
-
-#[derive(Debug)]
-pub struct ParityResult {
-    pub check_id: String,
-    pub batch_result: CheckResult,
-    pub streaming_result: CheckResult,
-    pub is_identical: bool,
-}
-
-impl ParityResult {
-    pub fn assert_parity(&self) {
-        if !self.is_identical {
-            panic!(
-                "PARITY VIOLATION for check '{}':\n\
-                 Batch:     {:?} - {}\n\
-                 Streaming: {:?} - {}\n\
-                 \n\
-                 This is a CRITICAL bug. Batch and streaming modes must produce identical results.",
-                self.check_id,
-                self.batch_result.outcome,
-                self.batch_result.reason,
-                self.streaming_result.outcome,
-                self.streaming_result.reason,
-            );
-        }
-    }
-}
-
-// ============================================================
-// Test Fixtures
-// ============================================================
-
-pub mod fixtures {
-    use super::*;
-
-    /// Generate a comprehensive set of test cases
-    pub fn all_test_cases() -> Vec<(PolicyCheck, CheckInput, Outcome)> {
-        let mut cases = Vec::new();
-
-        // ArgsValid test cases
-        cases.extend(args_valid_cases());
-        cases.extend(sequence_valid_cases());
-        cases.extend(blocklist_cases());
-        cases.extend(edge_cases());
-
-        cases
-    }
-
-    fn args_valid_cases() -> Vec<(PolicyCheck, CheckInput, Outcome)> {
-        let schema = serde_json::json!({
-            "type": "object",
-            "properties": {
-                "percent": { "type": "number", "maximum": 30 },
-                "reason": { "type": "string" }
-            },
-            "required": ["percent", "reason"]
-        });
-
-        vec![
-            // Pass: valid args
-            (
-                PolicyCheck {
-                    id: "args_valid_pass".into(),
-                    check_type: CheckType::ArgsValid,
-                    params: serde_json::json!({ "schema": schema }),
-                },
-                CheckInput {
-                    tool_name: Some("ApplyDiscount".into()),
-                    args: Some(serde_json::json!({
-                        "percent": 15,
-                        "reason": "Loyalty discount"
-                    })),
-                    trace: None,
-                },
-                Outcome::Pass,
-            ),
-            // Fail: exceeds maximum
-            (
-                PolicyCheck {
-                    id: "args_valid_exceed_max".into(),
-                    check_type: CheckType::ArgsValid,
-                    params: serde_json::json!({ "schema": schema }),
-                },
-                CheckInput {
-                    tool_name: Some("ApplyDiscount".into()),
-                    args: Some(serde_json::json!({
-                        "percent": 50,
-                        "reason": "Too much"
-                    })),
-                    trace: None,
-                },
-                Outcome::Fail,
-            ),
-            // Fail: missing required field
-            (
-                PolicyCheck {
-                    id: "args_valid_missing_required".into(),
-                    check_type: CheckType::ArgsValid,
-                    params: serde_json::json!({ "schema": schema }),
-                },
-                CheckInput {
-                    tool_name: Some("ApplyDiscount".into()),
-                    args: Some(serde_json::json!({
-                        "percent": 10
-                        // missing "reason"
-                    })),
-                    trace: None,
-                },
-                Outcome::Fail,
-            ),
-            // Error: no args provided
-            (
-                PolicyCheck {
-                    id: "args_valid_no_args".into(),
-                    check_type: CheckType::ArgsValid,
-                    params: serde_json::json!({ "schema": schema }),
-                },
-                CheckInput {
-                    tool_name: Some("ApplyDiscount".into()),
-                    args: None,
-                    trace: None,
-                },
-                Outcome::Error,
-            ),
-            // Edge: exactly at maximum (should pass)
-            (
-                PolicyCheck {
-                    id: "args_valid_at_max".into(),
-                    check_type: CheckType::ArgsValid,
-                    params: serde_json::json!({ "schema": schema }),
-                },
-                CheckInput {
-                    tool_name: Some("ApplyDiscount".into()),
-                    args: Some(serde_json::json!({
-                        "percent": 30,
-                        "reason": "Maximum allowed"
-                    })),
-                    trace: None,
-                },
-                Outcome::Pass,
-            ),
-        ]
-    }
-
-    fn sequence_valid_cases() -> Vec<(PolicyCheck, CheckInput, Outcome)> {
-        let rules = serde_json::json!({
-            "rules": [
-                { "type": "require", "tool": "VerifyIdentity" },
-                { "type": "before", "first": "VerifyIdentity", "then": "DeleteAccount" }
-            ]
-        });
-
-        vec![
-            // Pass: correct sequence
-            (
-                PolicyCheck {
-                    id: "sequence_valid_pass".into(),
-                    check_type: CheckType::SequenceValid,
-                    params: rules.clone(),
-                },
-                CheckInput {
-                    tool_name: None,
-                    args: None,
-                    trace: Some(vec![
-                        ToolCall {
-                            tool_name: "VerifyIdentity".into(),
-                            args: serde_json::json!({}),
-                            timestamp_ms: 1000,
-                        },
-                        ToolCall {
-                            tool_name: "ConfirmAction".into(),
-                            args: serde_json::json!({}),
-                            timestamp_ms: 2000,
-                        },
-                        ToolCall {
-                            tool_name: "DeleteAccount".into(),
-                            args: serde_json::json!({}),
-                            timestamp_ms: 3000,
-                        },
-                    ]),
-                },
-                Outcome::Pass,
-            ),
-            // Fail: missing required tool
-            (
-                PolicyCheck {
-                    id: "sequence_missing_required".into(),
-                    check_type: CheckType::SequenceValid,
-                    params: rules.clone(),
-                },
-                CheckInput {
-                    tool_name: None,
-                    args: None,
-                    trace: Some(vec![ToolCall {
-                        tool_name: "DeleteAccount".into(),
-                        args: serde_json::json!({}),
-                        timestamp_ms: 1000,
-                    }]),
-                },
-                Outcome::Fail,
-            ),
-            // Fail: wrong order
-            (
-                PolicyCheck {
-                    id: "sequence_wrong_order".into(),
-                    check_type: CheckType::SequenceValid,
-                    params: rules.clone(),
-                },
-                CheckInput {
-                    tool_name: None,
-                    args: None,
-                    trace: Some(vec![
-                        ToolCall {
-                            tool_name: "DeleteAccount".into(),
-                            args: serde_json::json!({}),
-                            timestamp_ms: 1000,
-                        },
-                        ToolCall {
-                            tool_name: "VerifyIdentity".into(),
-                            args: serde_json::json!({}),
-                            timestamp_ms: 2000,
-                        },
-                    ]),
-                },
-                Outcome::Fail,
-            ),
-            // Error: no trace
-            (
-                PolicyCheck {
-                    id: "sequence_no_trace".into(),
-                    check_type: CheckType::SequenceValid,
-                    params: rules.clone(),
-                },
-                CheckInput {
-                    tool_name: None,
-                    args: None,
-                    trace: None,
-                },
-                Outcome::Error,
-            ),
-        ]
-    }
-
-    fn blocklist_cases() -> Vec<(PolicyCheck, CheckInput, Outcome)> {
-        let params = serde_json::json!({
-            "blocked": ["DeleteDatabase", "DropTable", "ExecuteRawSQL"]
-        });
-
-        vec![
-            // Pass: allowed tool
-            (
-                PolicyCheck {
-                    id: "blocklist_allowed".into(),
-                    check_type: CheckType::ToolBlocklist,
-                    params: params.clone(),
-                },
-                CheckInput {
-                    tool_name: Some("SelectQuery".into()),
-                    args: None,
-                    trace: None,
-                },
-                Outcome::Pass,
-            ),
-            // Fail: blocked tool
-            (
-                PolicyCheck {
-                    id: "blocklist_blocked".into(),
-                    check_type: CheckType::ToolBlocklist,
-                    params: params.clone(),
-                },
-                CheckInput {
-                    tool_name: Some("DeleteDatabase".into()),
-                    args: None,
-                    trace: None,
-                },
-                Outcome::Fail,
-            ),
-            // Fail: another blocked tool
-            (
-                PolicyCheck {
-                    id: "blocklist_drop_table".into(),
-                    check_type: CheckType::ToolBlocklist,
-                    params: params.clone(),
-                },
-                CheckInput {
-                    tool_name: Some("DropTable".into()),
-                    args: None,
-                    trace: None,
-                },
-                Outcome::Fail,
-            ),
-            // Error: no tool name
-            (
-                PolicyCheck {
-                    id: "blocklist_no_tool".into(),
-                    check_type: CheckType::ToolBlocklist,
-                    params: params.clone(),
-                },
-                CheckInput {
-                    tool_name: None,
-                    args: None,
-                    trace: None,
-                },
-                Outcome::Error,
-            ),
-        ]
-    }
-
-    fn edge_cases() -> Vec<(PolicyCheck, CheckInput, Outcome)> {
-        let _schema = serde_json::json!({
-             "type": "string",
-             "minLength": 5
-        });
-
-        vec![
-            // 1. Empty args object where schema expects something
-            (
-                PolicyCheck {
-                    id: "edge_empty_args".into(),
-                    check_type: CheckType::ArgsValid,
-                    params: serde_json::json!({
-                        "schema": { "type": "object", "required": ["foo"] }
-                    }),
-                },
-                CheckInput {
-                    tool_name: Some("EdgeTool".into()),
-                    args: Some(serde_json::json!({})),
-                    trace: None,
-                },
-                Outcome::Fail,
-            ),
-            // 2. Null args where schema expects object
-            (
-                PolicyCheck {
-                    id: "edge_null_args".into(),
-                    check_type: CheckType::ArgsValid,
-                    params: serde_json::json!({
-                        "schema": { "type": "object" }
-                    }),
-                },
-                CheckInput {
-                    tool_name: Some("EdgeTool".into()),
-                    args: Some(serde_json::json!(null)),
-                    trace: None,
-                },
-                Outcome::Fail,
-            ),
-            // 3. Deeply nested schema
-            (
-                PolicyCheck {
-                    id: "edge_deep_nesting".into(),
-                    check_type: CheckType::ArgsValid,
-                    params: serde_json::json!({
-                        "schema": {
-                            "type": "object",
-                            "properties": {
-                                "a": { "type": "object", "properties": {
-                                    "b": { "type": "object", "properties": {
-                                        "c": { "type": "integer", "minimum": 0 }
-                                    }}
-                                }}
-                            }
-                        }
-                    }),
-                },
-                CheckInput {
-                    tool_name: Some("DeepTool".into()),
-                    args: Some(serde_json::json!({ "a": { "b": { "c": -1 } } })),
-                    trace: None,
-                },
-                Outcome::Fail,
-            ),
-            // 4. Unicode in tool name and args
-            (
-                PolicyCheck {
-                    id: "edge_unicode".into(),
-                    check_type: CheckType::ToolBlocklist,
-                    params: serde_json::json!({
-                        "blocked": ["🔥DangerousTool"]
-                    }),
-                },
-                CheckInput {
-                    tool_name: Some("🔥DangerousTool".into()),
-                    args: None,
-                    trace: None,
-                },
-                Outcome::Fail,
-            ),
-            // 5. Sequence with repeating tools
-            (
-                PolicyCheck {
-                    id: "edge_sequence_repeat".into(),
-                    check_type: CheckType::SequenceValid,
-                    params: serde_json::json!({
-                        "rules": [{ "type": "require", "tool": "Login" }]
-                    }),
-                },
-                CheckInput {
-                    tool_name: None,
-                    args: None,
-                    trace: Some(vec![
-                        ToolCall {
-                            tool_name: "Login".into(),
-                            args: serde_json::json!({}),
-                            timestamp_ms: 1,
-                        },
-                        ToolCall {
-                            tool_name: "Login".into(),
-                            args: serde_json::json!({}),
-                            timestamp_ms: 2,
-                        },
-                    ]),
-                },
-                Outcome::Pass,
-            ),
-            // 6. Huge number
-            (
-                PolicyCheck {
-                    id: "edge_huge_number".into(),
-                    check_type: CheckType::ArgsValid,
-                    params: serde_json::json!({
-                         "schema": { "type": "object", "properties": { "val": { "maximum": 100 } } }
-                    }),
-                },
-                CheckInput {
-                    tool_name: Some("MathTool".into()),
-                    args: Some(serde_json::json!({ "val": 1.0e+25 })),
-                    trace: None,
-                },
-                Outcome::Fail,
-            ),
-        ]
-    }
-}
-
-// ============================================================
-// Tests
-// ============================================================
-
+#[path = "parity/assertions.rs"]
+mod assertions;
+#[path = "parity/core_types.rs"]
+mod core_types;
 #[cfg(test)]
-mod tests {
-    use super::*;
+#[path = "parity/parity_contract.rs"]
+mod parity_contract;
 
-    #[test]
-    fn test_all_parity() {
-        let cases = fixtures::all_test_cases();
-        let mut failures = Vec::new();
+#[path = "parity/batch.rs"]
+pub mod batch;
+#[path = "parity/fixtures.rs"]
+pub mod fixtures;
+#[path = "parity/shared.rs"]
+pub mod shared;
+#[path = "parity/streaming.rs"]
+pub mod streaming;
 
-        println!("\n========================================");
-        println!("Parity Test: Batch vs Streaming");
-        println!("========================================\n");
-
-        for (check, input, expected_outcome) in &cases {
-            let result = verify_parity(check, input);
-
-            let status = if result.is_identical {
-                "✓ PARITY"
-            } else {
-                failures.push(result.check_id.clone());
-                "✗ DIVERGED"
-            };
-
-            let outcome_check = if result.batch_result.outcome == *expected_outcome {
-                "correct"
-            } else {
-                "WRONG OUTCOME"
-            };
-
-            println!(
-                "{} {} [{:?}] ({})",
-                status, check.id, result.batch_result.outcome, outcome_check
-            );
-        }
-
-        println!("\n----------------------------------------");
-        println!("Total: {} checks", cases.len());
-        println!("Parity: {} passed", cases.len() - failures.len());
-        println!("Diverged: {}", failures.len());
-        println!("----------------------------------------\n");
-
-        if !failures.is_empty() {
-            panic!(
-                "PARITY TEST FAILED\n\
-                 The following checks produced different results in batch vs streaming:\n\
-                 {:?}\n\n\
-                 This is a RELEASE BLOCKER.",
-                failures
-            );
-        }
-
-        println!("✓ All parity checks passed!\n");
-    }
-
-    #[test]
-    fn test_args_valid_parity() {
-        let check = PolicyCheck {
-            id: "discount_check".into(),
-            check_type: CheckType::ArgsValid,
-            params: serde_json::json!({
-                "schema": {
-                    "properties": {
-                        "percent": { "maximum": 30 }
-                    },
-                    "required": ["percent"]
-                }
-            }),
-        };
-
-        let input = CheckInput {
-            tool_name: Some("ApplyDiscount".into()),
-            args: Some(serde_json::json!({ "percent": 50 })),
-            trace: None,
-        };
-
-        let result = verify_parity(&check, &input);
-        result.assert_parity();
-        assert_eq!(result.batch_result.outcome, Outcome::Fail);
-    }
-
-    #[test]
-    fn test_sequence_parity() {
-        let check = PolicyCheck {
-            id: "verify_before_delete".into(),
-            check_type: CheckType::SequenceValid,
-            params: serde_json::json!({
-                "rules": [
-                    { "type": "before", "first": "Verify", "then": "Delete" }
-                ]
-            }),
-        };
-
-        let input = CheckInput {
-            tool_name: None,
-            args: None,
-            trace: Some(vec![
-                ToolCall {
-                    tool_name: "Verify".into(),
-                    args: serde_json::json!({}),
-                    timestamp_ms: 1000,
-                },
-                ToolCall {
-                    tool_name: "Delete".into(),
-                    args: serde_json::json!({}),
-                    timestamp_ms: 2000,
-                },
-            ]),
-        };
-
-        let result = verify_parity(&check, &input);
-        result.assert_parity();
-        assert_eq!(result.batch_result.outcome, Outcome::Pass);
-    }
-
-    #[test]
-    fn test_blocklist_parity() {
-        let check = PolicyCheck {
-            id: "no_delete_db".into(),
-            check_type: CheckType::ToolBlocklist,
-            params: serde_json::json!({
-                "blocked": ["DeleteDatabase"]
-            }),
-        };
-
-        // Test allowed
-        let input_allowed = CheckInput {
-            tool_name: Some("SelectQuery".into()),
-            args: None,
-            trace: None,
-        };
-
-        let result = verify_parity(&check, &input_allowed);
-        result.assert_parity();
-        assert_eq!(result.batch_result.outcome, Outcome::Pass);
-
-        // Test blocked
-        let input_blocked = CheckInput {
-            tool_name: Some("DeleteDatabase".into()),
-            args: None,
-            trace: None,
-        };
-
-        let result = verify_parity(&check, &input_blocked);
-        result.assert_parity();
-        assert_eq!(result.batch_result.outcome, Outcome::Fail);
-    }
-
-    #[test]
-    fn test_hash_determinism() {
-        // Verify that result hashes are deterministic
-        let check = PolicyCheck {
-            id: "hash_test".into(),
-            check_type: CheckType::ArgsValid,
-            params: serde_json::json!({ "schema": {} }),
-        };
-
-        let input = CheckInput {
-            tool_name: None,
-            args: Some(serde_json::json!({})),
-            trace: None,
-        };
-
-        let result1 = batch::evaluate(&check, &input);
-        let result2 = batch::evaluate(&check, &input);
-        let result3 = streaming::evaluate(&check, &input);
-
-        assert_eq!(result1.result_hash, result2.result_hash);
-        assert_eq!(result1.result_hash, result3.result_hash);
-    }
-}
+pub use assertions::{compute_result_hash, verify_parity, ParityResult};
+pub use core_types::{CheckInput, CheckResult, CheckType, Outcome, PolicyCheck, ToolCall};
